addressed feedback: no private API, shared -on-component util
