+++ addon/components/tooltip-on-component.js
@@ -1,20 +1,8 @@
-import Ember from 'ember';
 import TooltipOnElementComponent from 'ember-tooltips/components/tooltip-on-element';
-
-const { computed } = Ember;
+import { onComponentTarget } from 'ember-tooltips/utils';
 
 export default TooltipOnElementComponent.extend({
 
-  target: computed(function() {
-    const parentView = this.get('parentView');
-
-    if (!parentView) {
-      console.warn('No parentView found');
-
-      return null;
-    } else {
-      return `#${parentView.get('elementId')}`;
-    }
-  }),
+	target: onComponentTarget,
 
 });
